Save and load reminders to disk
This fixes data loss when restarting server. Some reminders will still be lost if they were supposed to be happen when the server was offline *during* the downtime.

Also currently the restored reminders from disk will only be printed when they're due, they are not sent as a direct message to the user.